# swagger.yml
openapi: 3.0.1
info:
  title: ThinkLab Manager API
  description: REST API for managing system users and related operations.
  contact:
    name: Development Team
    email: dev@thinklab.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: "0.1"
...
paths:
...
  /compute:
    post:
      tags:
      - Compute Resource Management
      summary: Register a new hybrid compute resource node
      description: "Logs and provisions a target cloud instance, on-prem VM, or physical\
        \ hardware asset profile."
      operationId: registerComputeNode
      parameters:
      - name: X-Executor
        in: header
        description: Identifier of the administrative engineer cataloging the asset
          node
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComputeResourceRequest"
        required: true
      responses:
        "201":
          description: Compute resource instance cataloged successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComputeResourceResponse"
        "400":
          description: Invalid payload constraints or unrecognized type configuration
            tokens
  /compute/{resourceHash}:
    get:
      tags:
      - Compute Resource Management
      summary: Retrieve a distinct compute profile configuration
      description: "Fetches complete specifications, active network layout mapping,\
        \ and the latest hardware usage telemetry snapshot."
      operationId: getComputeNodeByHash
      parameters:
      - name: resourceHash
        in: path
        description: The specific infra business key tracking token hash
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Compute node profile located and returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComputeResourceResponse"
        "404":
          description: Target resource tracking hash token not found in cluster storage
  /compute/{resourceHash}/maintenance/end:
    post:
      tags:
      - Compute Resource Management
      summary: Terminate maintenance mode and start verification loop
      description: Closes actual timestamps and returns node to PROVISIONING status
        to await clean agent telemetry synchronization.
      operationId: concludeActiveMaintenance
      parameters:
      - name: resourceHash
        in: path
        required: true
        schema:
          type: string
      - name: X-Executor
        in: header
        required: true
        schema:
          minLength: 1
          type: string
      responses:
        "200":
          description: Asset released from maintenance and routed back to verification
            queues
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComputeResourceResponse"
        "400":
          description: Target asset is not currently flagged under an open maintenance
            loop
        "404":
          description: Target compute resource hash not found
  /compute/{resourceHash}/maintenance/schedule:
    post:
      tags:
      - Compute Resource Management
      summary: Schedule a future maintenance window
      description: "Reserves a timeline block for physical or virtual hardware interventions,\
        \ mapping out expected duration and alarm silences."
      operationId: scheduleMaintenanceBlock
      parameters:
      - name: resourceHash
        in: path
        required: true
        schema:
          type: string
      - name: X-Executor
        in: header
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComputeResourceService.MaintenanceScheduleCmd"
        required: true
      responses:
        "200":
          description: Maintenance window successfully reserved in inventory timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComputeResourceResponse"
        "400":
          description: Invalid timeline boundaries or date execution logic sequence
            constraints
        "404":
          description: Target compute resource hash not found
  /compute/{resourceHash}/maintenance/start:
    post:
      tags:
      - Compute Resource Management
      summary: Activate maintenance isolation protocol
      description: "Shifts node into MAINTENANCE state, applying hard suppressions\
        \ against monitoring agent ping interference."
      operationId: initiateActiveMaintenance
      parameters:
      - name: resourceHash
        in: path
        required: true
        schema:
          type: string
      - name: X-Executor
        in: header
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComputeResourceService.MaintenanceStartCmd"
        required: true
      responses:
        "200":
          description: Asset locked and isolated under hard maintenance mode
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComputeResourceResponse"
        "404":
          description: Target compute resource hash not found
  /compute/{resourceHash}/telemetry:
    patch:
      tags:
      - Compute Resource Management
      summary: Ingest remote agent streaming telemetry metrics
      description: "Updates active processing strain load percentages, network throughput,\
        \ and overrides the node operational health status."
      operationId: ingestMetrics
      parameters:
      - name: resourceHash
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComputeResourceTelemetryRequest"
        required: true
      responses:
        "200":
          description: Telemetry packet successfully processed and counters overridden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComputeResourceResponse"
        "400":
          description: Invalid telemetry counters or parsing failure exception triggered
        "404":
          description: Target resource tracking hash not found to map metrics link
...
  /maintenance:
    post:
      tags:
      - Maintenance Workflow
      summary: Open a maintenance ticket
      description: Initializes a central tracking record bound against an existing
        inventory hardware asset hash.
      operationId: openTicket
      parameters:
      - name: X-Executor
        in: header
        description: Identifier of the agent dispatching the pipeline action
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaintenanceTicketRequest"
        required: true
      responses:
        "201":
          description: Ticket successfully opened and cataloged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceTicketResponse"
        "400":
          description: Invalid payload structural attributes
        "404":
          description: Target infrastructure asset hash not found
  /maintenance/{ticketHash}:
    get:
      tags:
      - Maintenance Workflow
      summary: Retrieve a ticket profile
      description: "Fetches complete ongoing details, timeline commentary logs, and\
        \ status scopes of a ticket."
      operationId: getTicketByHash
      parameters:
      - name: ticketHash
        in: path
        description: The specific business token string hash reference
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Maintenance tracking ticket found and returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceTicketResponse"
        "404":
          description: Ticket not found matching hash parameter
  /maintenance/{ticketHash}/comments:
    post:
      tags:
      - Maintenance Workflow
      summary: Add a progress comment
      description: Appends a historical immutable textual diagnostic or status notation
        log directly inside the document timeline graph.
      operationId: appendComment
      parameters:
      - name: ticketHash
        in: path
        required: true
        schema:
          type: string
      - name: X-Executor
        in: header
        description: Identifier of the human technician or webhook updating the queue
          trace
        required: true
        schema:
          minLength: 1
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCommentRequest"
        required: true
      responses:
        "200":
          description: Comment successfully appended into the timeline list document
            graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceTicketResponse"
        "404":
          description: Ticket target hash parameter reference not found
  /maintenance/{ticketHash}/status:
    patch:
      tags:
      - Maintenance Workflow
      summary: Transition ticket lifecycle state
      description: Alters the internal operational track enum status location nodes
        using strict guardrails checking mechanisms.
      operationId: transitionStatus
      parameters:
      - name: ticketHash
        in: path
        required: true
        schema:
          type: string
      - name: targetStatus
        in: query
        description: The raw target MaintenanceStatus string node token enum
        required: true
        explode: false
        schema:
          type: string
      - name: X-Executor
        in: header
        description: Identifier of the supervisor executing systemic lifecycle override
        required: true
        schema:
          minLength: 1
          type: string
      responses:
        "200":
          description: Status updated and saved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaintenanceTicketResponse"
        "400":
          description: State validation guard check bounds violation or terminal node
            block lock
...
components:
  schemas:
    AddCommentRequest:
      required:
      - author
      - message
      type: object
      properties:
        author:
          minLength: 1
          type: string
          x-size-message: Comment author cannot be blank
        message:
          minLength: 1
          type: string
          x-size-message: Comment message content text cannot be blank
      description: Inbound payload representing an atomic discursive addition inside
        an active maintenance pipeline.
...
    ComputeResourceRequest:
      required:
      - externalFqdn
      - gateway
      - internalDnsName
      - macAddress
      - name
      - operatingSystem
      - privateIp
      - provider
      - publicIp
      - ramMemoryMb
      - resourceType
      - spec
      - storageCapacityGb
      - subnetCidr
      - vCpus
      type: object
      properties:
        name:
          minLength: 1
          type: string
          x-size-message: Resource instance descriptive name cannot be blank
        resourceType:
          minLength: 1
          type: string
          x-size-message: Resource virtualization architecture type layer cannot be
            blank
        provider:
          minLength: 1
          type: string
          x-size-message: Infrastructure hosting provider deployment location cannot
            be blank
        vCpus:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int32
          x-minimum-message: vCPUs volume allocation must be greater than zero
          x-not-null-message: Allocated virtual CPU core quantity must be provided
        ramMemoryMb:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int32
          x-minimum-message: RAM memory allocation size in Megabytes must be greater
            than zero
          x-not-null-message: Allocated RAM memory size must be provided
        storageCapacityGb:
          minimum: 0
          exclusiveMinimum: true
          type: integer
          format: int64
          x-minimum-message: Storage capacity allocation in Gigabytes must be greater
            than zero
          x-not-null-message: Persistent storage block volume capacity must be provided
        operatingSystem:
          minLength: 1
          type: string
          x-size-message: Target node operating system or base image template name
            cannot be blank
        privateIp:
          type: string
        publicIp:
          type: string
        subnetCidr:
          type: string
        gateway:
          type: string
        macAddress:
          type: string
        internalDnsName:
          type: string
        externalFqdn:
          type: string
        spec:
          type: object
          additionalProperties:
            type: object
      description: "Inbound API Request payload containing comprehensive parameters\
        \ for registering, provisioning, and cataloging a cloud or on-premises compute\
        \ resource."
    ComputeResourceResponse:
      required:
      - createdAt
      - maintenanceWindow
      - metrics
      - name
      - networkConfig
      - operatingSystem
      - provider
      - ramMemoryMb
      - resourceHash
      - resourceType
      - spec
      - status
      - storageCapacityGb
      - updatedAt
      - vCpus
      type: object
      properties:
        resourceHash:
          type: string
        name:
          type: string
        resourceType:
          $ref: "#/components/schemas/ResourceType"
        provider:
          $ref: "#/components/schemas/InfrastructureProvider"
        status:
          $ref: "#/components/schemas/ResourceStatus"
        vCpus:
          type: integer
          format: int32
        ramMemoryMb:
          type: integer
          format: int32
        storageCapacityGb:
          type: integer
          format: int64
        operatingSystem:
          type: string
        networkConfig:
          $ref: "#/components/schemas/NetworkConfig"
        metrics:
          $ref: "#/components/schemas/ResourceMetrics"
        maintenanceWindow:
          $ref: "#/components/schemas/MaintenanceWindow"
        spec:
          type: object
          additionalProperties:
            type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      description: "Outbound unified JSON data snapshot representing a complete hybrid,\
        \ cloud, or on-premises compute environment profile with active network and\
        \ SRE maintenance configurations."
    ComputeResourceService.MaintenanceScheduleCmd:
      required:
      - reason
      - scheduledEnd
      - scheduledStart
      - silenceAlerts
      type: object
      properties:
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        reason:
          minLength: 1
          type: string
        silenceAlerts:
          type: boolean
      description: Inbound command record DTO dedicated to scheduling future resource
        maintenance windows.
    ComputeResourceService.MaintenanceStartCmd:
      required:
      - reason
      - silenceAlerts
      type: object
      properties:
        reason:
          minLength: 1
          type: string
        silenceAlerts:
          type: boolean
      description: Inbound command record DTO dedicated to starting an immediate or
        scheduled maintenance window.
    ComputeResourceTelemetryRequest:
      required:
      - cpuUsagePercentage
      - diskUsageGb
      - networkInMbitSec
      - networkOutMbitSec
      - ramUsageMb
      - status
      type: object
      properties:
        status:
          minLength: 1
          type: string
          x-size-message: Operational status signature node cannot be blank
        cpuUsagePercentage:
          maximum: 100
          minimum: 0
          type: number
          format: double
          x-maximum-message: CPU percentage consumption ceiling bounds cannot exceed
            100%
          x-minimum-message: CPU percentage consumption floor bounds must be 0%
          x-not-null-message: Instantaneous CPU usage ratio core must be provided
        ramUsageMb:
          minimum: 0
          type: integer
          format: int64
          x-minimum-message: RAM memory utilization size in MB cannot be negative
          x-not-null-message: Active RAM memory consumption metrics must be provided
        diskUsageGb:
          minimum: 0
          type: integer
          format: int64
          x-minimum-message: Storage utilization size in GB cannot be negative
          x-not-null-message: Active disk space consumption metrics must be provided
        networkInMbitSec:
          type: number
          format: double
        networkOutMbitSec:
          type: number
          format: double
      description: "High-performance inbound telemetry payload pushed by remote monitoring\
        \ agents or cloud metrics scrapers. Updates dynamic strain levels, capacity\
        \ usage, and overrides heartbeat states."
...
    InfrastructureProvider:
      type: string
      description: "Defines the structural hosting architecture zones for the compute\
        \ infrastructure. Maps high-availability public clouds, bare-metal enterprise\
        \ datacenters, and edge home-lab environments."
      enum:
      - AWS_CLOUD
      - AZURE_CLOUD
      - GCP_CLOUD
      - PRIVATE_DATACENTER
      - HOME_LAB
      x-enum-descriptions:
      - Amazon Web Services elastic compute global public cloud infrastructure.
      - Microsoft Azure public enterprise cloud computing matrix.
      - Google Cloud Platform data-driven hyper-scaler infrastructure network.
      - "Private on-premises Tier-3/Tier-4 physical datacenter environments (Datacenter\
        \ X, Datacenter Y)."
      - Localized edge home-lab deployment ecosystem running behind custom proxy blocks
        or residential networks.
...
    MaintenanceTicketRequest:
      required:
      - assetHash
      - assignedTechnicianId
      - issueDescription
      - requesterId
      type: object
      properties:
        assetHash:
          minLength: 1
          type: string
          x-size-message: Target asset hash reference cannot be blank
        requesterId:
          minLength: 1
          type: string
          x-size-message: Requester identifier cannot be blank
        assignedTechnicianId:
          type: string
        issueDescription:
          minLength: 1
          type: string
          x-size-message: Detailed issue description cannot be blank
      description: API Request payload bound to input validations for launching a
        new hardware maintenance ticket.
    MaintenanceTicketResponse:
      required:
      - assetHash
      - assignedTechnicianId
      - comments
      - createdAt
      - issueDescription
      - requesterId
      - status
      - ticketHash
      - updatedAt
      type: object
      properties:
        ticketHash:
          type: string
        assetHash:
          type: string
        requesterId:
          type: string
        assignedTechnicianId:
          type: string
        issueDescription:
          type: string
        status:
          $ref: "#/components/schemas/MaintenanceStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        comments:
          type: array
          items:
            $ref: "#/components/schemas/MaintenanceComment"
      description: Outbound complete structural JSON snapshot representing a corporate
        maintenance ticket instance.
...
    NetworkConfig:
      required:
      - externalFqdn
      - gateway
      - internalDnsName
      - macAddress
      - privateIp
      - publicIp
      - subnetCidr
      type: object
      properties:
        privateIp:
          type: string
        publicIp:
          type: string
        subnetCidr:
          type: string
        gateway:
          type: string
        macAddress:
          type: string
        internalDnsName:
          type: string
        externalFqdn:
          type: string
      description: "Embedded subdocument tracking detailed network topology, routing\
        \ vectors, hardware mac identification, and DNS configurations across public\
        \ or private segments."
    ResourceMetrics:
      required:
      - cpuUsagePercentage
      - diskUsageGb
      - lastTelemetryAt
      - networkInMbitSec
      - networkOutMbitSec
      - ramUsageMb
      type: object
      properties:
        cpuUsagePercentage:
          type: number
          format: double
        ramUsageMb:
          type: integer
          format: int64
        diskUsageGb:
          type: integer
          format: int64
        networkInMbitSec:
          type: number
          format: double
        networkOutMbitSec:
          type: number
          format: double
        lastTelemetryAt:
          type: string
          format: date-time
      description: "Embedded subdocument tracking real-time telemetry, capacity thresholds,\
        \ and hardware processing strain signatures for a compute resource."
...
    ResourceType:
      type: string
      description: Defines the structural compute virtualization level and execution
        architecture layer.
      enum:
      - BARE_METAL_HOST
      - VIRTUAL_MACHINE
      - CONTAINER
      x-enum-descriptions:
      - Bare-metal physical enterprise hardware server rack unit hosting hypervisors
        or cluster engines.
      - "Traditional hypervisor-managed abstraction layer instance (e.g., AWS EC2,\
        \ Proxmox VM, VMware ESXi)."
      - "Lightweight OS-level virtualized isolate processing node (e.g., Docker Container,\
        \ Podman Core, K8s Pod)."
...